Add notes file
The workflow I will follow is that I will define the first two
steps of the user interface flow, and the corresponding
components that need to be built out.

Once that is complete and working, I will go back
and update the notes with the subsequent two steps.

This is open to modification as I progress through
building this app.