Correct form for new teacher and add views
Remakes the teacher form. Previously had used
rails generate controller action action, to make the controller.
This was yielding an error, because the @teacher instance variable in
teachers#new was not passing on to the new.html.erb view file.

When redone using rails generate controller without specifying
controller actions it works. Manually constructed contents
of controller as well as created the view files.

This is a working flow for a new teacher to sign up,
without authentication.